requirements.txt
streamlit
python-dotenv
psycopg2-binary
mysql-connector-python
pymongo
google-generativeai
